+++ packages/sdks-tests/src/e2e-tests/repeat-items-bindings.spec.ts
@@ -15,4 +15,17 @@ test.describe('Repeat items bindings', () => {
     await expect(page.locator('text=4')).toBeVisible();
     await expect(page.locator('text=5')).toBeVisible();
   });
+
+  test('repeat Symbols with content bound to content inputs should render correctly', async ({
+    page,
+  }) => {
+    // here data prop is { products: [{ header: 'title1' }, { header: 'title2' }, { header: 'title3' }] }
+    await page.goto('/symbol-with-repeat-input-binding');
+
+    const promises = [];
+    for (let i = 1; i <= 3; i++) {
+      promises.push(expect(page.locator(`text=title${i}`)).toBeVisible());
+    }
+    await Promise.all(promises);
+  });
 });